Mem usage optimization

+++ model/din/activation_test.go
@@ -11,8 +11,8 @@ import (
 func TestPRelu(t *testing.T) {
 	Convey("prelu", t, func() {
 		g := G.NewGraph()
-		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(4, 1), tensor.WithBacking([]float64{-1, -2, 3, 4})), G.WithName("x"))
-		a := G.NewScalar(g, G.Float64, G.WithValue(0.1), G.WithName("a"))
+		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(4, 1), tensor.WithBacking([]float32{-1, -2, 3, 4})), G.WithName("x"))
+		a := G.NewScalar(g, G.Float32, G.WithValue(0.1), G.WithName("a"))
 		output := PRelu(x, a)
 		//cost := G.Must(G.Mean(output))
 		//
@@ -24,15 +24,15 @@ func TestPRelu(t *testing.T) {
 			t.Fatalf("%+v", err)
 		}
 		defer m.Close()
-		So(output.Value().Data(), ShouldResemble, []float64{-0.1, -0.2, 3, 4})
+		So(output.Value().Data(), ShouldResemble, []float32{-0.1, -0.2, 3, 4})
 	})
 }
 
 func TestEucDistance(t *testing.T) {
 	Convey("euc distance 2 dim", t, func() {
 		g := G.NewGraph()
-		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(2, 4), tensor.WithBacking([]float64{1, 2, -3, 4, -1, 0, -1, 2})), G.WithName("x"))
-		y := G.NodeFromAny(g, tensor.New(tensor.WithShape(2, 4), tensor.WithBacking([]float64{1, 2, -3, 4, 0, 1, 0, 1})), G.WithName("y"))
+		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(2, 4), tensor.WithBacking([]float32{1, 2, -3, 4, -1, 0, -1, 2})), G.WithName("x"))
+		y := G.NodeFromAny(g, tensor.New(tensor.WithShape(2, 4), tensor.WithBacking([]float32{1, 2, -3, 4, 0, 1, 0, 1})), G.WithName("y"))
 		output := EucDistance(x, y)
 		m := G.NewTapeMachine(g)
 		if err := m.RunAll(); err != nil {
@@ -40,12 +40,12 @@ func TestEucDistance(t *testing.T) {
 		}
 		defer m.Close()
 		So([]int(output.Shape()), ShouldResemble, []int{2})
-		So(output.Value().Data(), ShouldResemble, []float64{0, 2})
+		So(output.Value().Data(), ShouldResemble, []float32{0, 2})
 	})
 	Convey("euc distance 3 dim no broadcast", t, func() {
 		g := G.NewGraph()
-		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 2, 2), tensor.WithBacking([]float64{1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0})), G.WithName("x"))
-		y := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 2, 2), tensor.WithBacking([]float64{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0})), G.WithName("y"))
+		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 2, 2), tensor.WithBacking([]float32{1, 0, 0, 0, 1, 0, 1, 0, 0, 0, 1, 0})), G.WithName("x"))
+		y := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 2, 2), tensor.WithBacking([]float32{0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0})), G.WithName("y"))
 		output := EucDistance(x, y)
 		m := G.NewTapeMachine(g)
 		if err := m.RunAll(); err != nil {
@@ -53,12 +53,12 @@ func TestEucDistance(t *testing.T) {
 		}
 		defer m.Close()
 		So([]int(output.Shape()), ShouldResemble, []int{3, 2})
-		So(output.Value().Data(), ShouldResemble, []float64{1, 0, 1, 1, 0, 1})
+		So(output.Value().Data(), ShouldResemble, []float32{1, 0, 1, 1, 0, 1})
 	})
 	Convey("euc distance 3 dim broadcast y", t, func() {
 		g := G.NewGraph()
-		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 2, 2), tensor.WithBacking([]float64{0, 2, 0, 0, -1, 1, -1, 1, 1, 2, 2, 2})), G.WithName("x"))
-		y := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 1, 2), tensor.WithBacking([]float64{0, 1, 0, 1, 1, 2})), G.WithName("y"))
+		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 2, 2), tensor.WithBacking([]float32{0, 2, 0, 0, -1, 1, -1, 1, 1, 2, 2, 2})), G.WithName("x"))
+		y := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 1, 2), tensor.WithBacking([]float32{0, 1, 0, 1, 1, 2})), G.WithName("y"))
 		output := EucDistance(x, y)
 		m := G.NewTapeMachine(g)
 		if err := m.RunAll(); err != nil {
@@ -66,12 +66,12 @@ func TestEucDistance(t *testing.T) {
 		}
 		defer m.Close()
 		So([]int(output.Shape()), ShouldResemble, []int{3, 2})
-		So(output.Value().Data(), ShouldResemble, []float64{1, 1, 1, 1, 0, 1})
+		So(output.Value().Data(), ShouldResemble, []float32{1, 1, 1, 1, 0, 1})
 	})
 	Convey("euc distance 3 dim broadcast x", t, func() {
 		g := G.NewGraph()
-		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 1, 2), tensor.WithBacking([]float64{0, 1, 0, 1, 1, 2})), G.WithName("x"))
-		y := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 2, 2), tensor.WithBacking([]float64{0, 2, 0, 0, -1, 1, -1, 1, 1, 2, 2, 2})), G.WithName("y"))
+		x := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 1, 2), tensor.WithBacking([]float32{0, 1, 0, 1, 1, 2})), G.WithName("x"))
+		y := G.NodeFromAny(g, tensor.New(tensor.WithShape(3, 2, 2), tensor.WithBacking([]float32{0, 2, 0, 0, -1, 1, -1, 1, 1, 2, 2, 2})), G.WithName("y"))
 		output := EucDistance(x, y)
 		m := G.NewTapeMachine(g)
 		if err := m.RunAll(); err != nil {
@@ -79,6 +79,6 @@ func TestEucDistance(t *testing.T) {
 		}
 		defer m.Close()
 		So([]int(output.Shape()), ShouldResemble, []int{3, 2})
-		So(output.Value().Data(), ShouldResemble, []float64{1, 1, 1, 1, 0, 1})
+		So(output.Value().Data(), ShouldResemble, []float32{1, 1, 1, 1, 0, 1})
 	})
 }
